+++ src/lib/LangChoose.svelte
@@ -1,10 +1,10 @@
 <script lang="ts">
-    import LangChooseOption from "./LangChooseOption.svelte";
+    import LangOption from "./LangOption.svelte";
     export let lang1Option:{id:number, language:string, print:string, integer:string, string:string, decimal:string};
     export let lang2Option:{id:number, language:string, print:string, integer:string, string:string, decimal:string};
 </script>
 
 <div>
-    Compare <LangChooseOption  bind:chosenLanguage={lang1Option}/> 
-    syntax with <LangChooseOption bind:chosenLanguage={lang2Option}/> syntax
+    Compare <LangOption  bind:chosenLanguage={lang1Option}/> 
+    syntax with <LangOption bind:chosenLanguage={lang2Option}/> syntax
 </div>
